requirements.txt
openai==0.26.5
python-dotenv==1.0.0
mock==5.0.1
google-auth-oauthlib==1.0.0
google-api-python-client==2.79.0
Flask==2.2.3
google-auth==2.16.1
pytest==7.2.1